Put position check into a phase
